Code Cleanup
Finalization of the Engine itself( for the most part) Code was cleaned up, Enemies, their projectiles, the player and his projectiles all spawn/Despawn effortlessly. Sound is working although music is a little low, There is a slight memory leak possibly with the projectiles. 
Time for Asset management and GameDesign.